LayoutTests/imported/w3c/web-platform-tests/css/selectors/invalidation/host-has-shadow-tree-element-at-nonsubject-position-expected.txt


FAIL Initial color assert_equals: expected "rgb(255, 0, 0)" but got "rgb(255, 192, 203)"
FAIL Add .descendant to #shadow_child assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 192, 203)"
FAIL Remove .descendant from #shadow_child assert_equals: expected "rgb(255, 0, 0)" but got "rgb(255, 192, 203)"
FAIL Add .descendant to #shadow_descendant assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 192, 203)"
FAIL Add .ancestor to #shadow_child:has(.descendant) assert_equals: expected "rgb(128, 0, 128)" but got "rgb(255, 192, 203)"
FAIL Remove .ancestor from #shadow_child:has(.descendant) assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 192, 203)"
PASS Add .child to #shadow_child:has(.descendant)
FAIL Remove .child from #shadow_child:has(.descendant) assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 192, 203)"
FAIL Remove .descendant from #shadow_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(255, 192, 203)"
FAIL Add .child to #shadow_child assert_equals: expected "rgb(0, 0, 255)" but got "rgb(255, 192, 203)"
FAIL Add .grand_child to #shadow_descendant assert_equals: expected "rgb(255, 255, 240)" but got "rgb(255, 192, 203)"
FAIL Add .host_context to #host assert_equals: expected "rgb(135, 206, 235)" but got "rgb(255, 192, 203)"
FAIL Add .descendant to #shadow_descendant.grand_child assert_equals: expected "rgb(144, 238, 144)" but got "rgb(255, 192, 203)"
FAIL Remove .descendant from #shadow_descendant.grand_child assert_equals: expected "rgb(135, 206, 235)" but got "rgb(255, 192, 203)"
FAIL Remove .grand_child from #shadow_descendant assert_equals: expected "rgb(0, 0, 255)" but got "rgb(255, 192, 203)"
FAIL Remove .child from #shadow_child assert_equals: expected "rgb(255, 0, 0)" but got "rgb(255, 192, 203)"
FAIL Add .child to #shadow_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(255, 192, 203)"
FAIL Remove .child from #shadow_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(255, 192, 203)"
FAIL Insert #first_child.descendant to shadow root assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 192, 203)"
FAIL Remove #first_child.descendant from shadow root assert_equals: expected "rgb(255, 0, 0)" but got "rgb(255, 192, 203)"
FAIL Insert #last_child.descendant to shadow root assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 192, 203)"
FAIL Remove #last_child.descendant from shadow root assert_equals: expected "rgb(255, 0, 0)" but got "rgb(255, 192, 203)"
FAIL Insert #child_in_middle.descendant before #shadow_child assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 192, 203)"
FAIL Remove #child_in_middle.descendant from shadow root assert_equals: expected "rgb(255, 0, 0)" but got "rgb(255, 192, 203)"
FAIL Insert #grand_child.descendant before #shadow_descendant assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 192, 203)"
FAIL Remove #grand_child.descendant from shadow tree assert_equals: expected "rgb(255, 0, 0)" but got "rgb(255, 192, 203)"
